remove try except in add_fnod

+++ pylayers/signal/bsignal.py
@@ -8,7 +8,7 @@ r"""
 Bsignal Class
 =============
 
-.. autoclass:: Bsignal 
+.. autoclass:: Bsignal
     :members:
 
 Usignal Class
@@ -3481,12 +3481,12 @@ class FHsignal(FUsignal):
         N = len(self.x)
         # even case
         if np.mod(N, 2) == 0:
-            xu = self.x[1:(N + 2) / 2]
-            yu = self.y[...,1:(N + 2) / 2]
+            xu = self.x[1:int((N + 2) / 2)]
+            yu = self.y[...,1:int((N + 2) / 2)]
         # odd case
         else:
-            xu = self.x[1:(N + 1) / 2]
-            yu = self.y[...,1:(N + 1) / 2]
+            xu = self.x[1:int((N + 1) / 2)]
+            yu = self.y[...,1:int((N + 1) / 2)]
 
         O = FUsignal(x=xu, y=yu)
